backend/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
django-environ==0.11.2
django-cors-headers==4.4.0
django-filter==24.3
psycopg[binary]==3.2.3
celery==5.4.0
redis==5.1.0
structlog==24.2.0
requests==2.32.3
weasyprint==61.2
reportlab==4.2.0
pyswisseph==2.10.3.2; python_version < "3.13"

